Backup tree structure.
lib/geo.php:build_input_html creates an HTML tree by iterating
over all base maps. But a tree structure appears to be very
inefficient. A lot of instructions have to be processed and
a parent-parents without children cannot be determined easily.
The old tree structure looks like this:

> Länder
  > Deutschland
  > Österreich
    > Föderale Ebene
      > Bundesländer Österreichs

A new flat structure like this is introduced:

> Länder / Deutschland
> Länder / Österreich
> Länder / Österreich / Föderale Ebene / Bundesländer Österreichs

A file_exists check can be added easily; a missing feature of the
structure above.